add: `Promise` getters to `KurtResult`, for convenience.
This will be more convenient for callers who are not interested
in the stream, and only want the final event or text/data from it.

This commit also adds a fair bit complexity to KurtResult,
so that it will behave correctly if there are multiple concurrent
"awaiters" using the `AsyncIterable` interface and/or the
convenience `Promise` getters.

Tests are included to ensure that we get the correct behavior
in many concurrent cases.