refactor: Shrink interface of KurtEvent and rename as KurtStream
As discussed with @InfraK, we want to shrink the interface a bit
because it will lessen the impact of future changes to the final
event structure.

In doing so, we also decided to rename some things, such that
the word `result` is only used for the final event of the stream,
rather than the entire stream itself.

BREAKING CHANGE: `finalText` and `finalData` have been removed, and some types have been renamed