refactor: change Kurt from an interface to an abstract class
This will enable Kurt to provide some convenience methods in a common place (implemented as a
composite of the abstract methods that each concrete class must implement.

An example of the kind of convenience method I'm talking about is described in [this
comment](https://github.com/FormulaMonks/kurt/pull/12#issuecomment-2099251542).

Overall, this is a breaking change so I want to introduce it before the impending public release.

BREAKING CHANGE: Kurt is now an abstract class.