refactor: make Kurt a concrete class that accepts an LLM-specific adapter object
Based on some discussion in Slack, we want to move away from the abstract class pattern for Kurt,
and also try to centralize some of the common elements across different sub-libraries.

This commit moves us in that direction by using an adapter object with functions that get composed
by the central Kurt class.

Note that we want to be able to evolve the adapter interface over time without hard breaking
changes, so I have set this up in a KurtAdapterV1 interface. A map of type parameters is used to
ensure that the intermediate data structures can have strong typing.