ci: add automatic releases with semantic-release-monorepo
This commit adds `semantic-release-monorepo` into our CI workflow,
which should automatically bump package versions and publish
releases for us, based on conventional commits.

Note that it is smart enough to notice when it's not on the main
branch, and in such situations it will skip the release process.
So there's no need for a branch guard in our CI YAML.

I have done some preliminary testing, but a full test won't
happen until we merge it and give it a try.

I can always revert if it doesn't work out well for us.