fix: make the adapters depend on non-workspace Kurt version
Prior to this commit, the adapter libraries were depending on a workspace version of Kurt. This is
unfit for publishing, because downstream users of the libraries won't have Kurt in their workspace.

We need to depend on fixed versions of Kurt and then use pnpm overrides to make it use the workspace
version of Kurt in local development.

This commit does that fix.