chore: remove `version` number from `package.json` for each library
We are using `semantic-release` which overrides the number in `package.json`
with an appropriate version number and doesn't commit it back to git.
To understand the reasons why it doesn't, please see this FAQ entry:
https://github.com/semantic-release/semantic-release/blob/73bcd0bcdb98322c5c5224ffb55960de114c3fd1/docs/support/FAQ.md#why-is-the-packagejsons-version-not-updated-in-my-repository

But having the `version` numbers in our repository is potentially confusing
because people may look to that field to see what the latest version is.

So this commit updates our repo to use the value `1.0.0-semantically-released`,
as per the recommendation of that same FAQ entry (but using `1` as the major
version so as not to break `pnpm` tooling), so we can use a less confusing
value that points to the right way to see our latest version, while still
technically adhering to the requirements of the `package.json` format
about a version number being present and using a valid version format.