feat: add `KurtSamplingOptions`, for tweaking output sampling details
This commit adds a `KurtSamplingOptions` type with various tweakable values and a
`KurtSamplingOptionsDefault` constant with the default values to use when not given.

The `sampling` key can be used to specify sampling options on creation, and/or on individual
generate calls. See the docs and tests for more details on what settings are possible and how they
behave.

A future PR will update the adapter libraries to start making use of these values.