refactor: update test suite to use automatic YAML snapshots
Adding new test cases was previously a bit of a painful
manual tasks involving a dance of temporarily adding some
`console.log` statements and setting `USE_REAL_API = true`,
then doing a lot of copy-pasting stuff around.

This commit removes that style of end-to-end testing in favor
automatic snapshotting of request, response, and processed
data in a YAML file that can be saved and loaded to replay
saved response data for mocking, as well as asserted against
as a snapshot specifying the correct request data and processed data.

This all happens automatically when you add a new test
with a unique name, using the `snapshotAndMock` method.

This automatic operation makes it very easy to add new tests,
therefore I consider it to be a valuable first step toward
increasing our test surface and testing more cases.

Note that the `snapshotAndMock` method is implemented
separately for `KurtOpenAI` and `KurtVertexAI` because
the implementation was different enough to cast doubt
on the value of unifying the two. But this is a decision
we can potentially revisit later as we add more plugins.