feat: pass KurtSamplingOptions to the underlying LLM provider
This is a followup to PR #24, which added `KurtSamplingOptions` to the `Kurt` class.

Now, with this PR, the LLM-specific adapters (`KurtOpenAI` and `KurtVertexAI`) will now pass these
options along to the underlying LLM provider, in the appropriate adapted way.

This commit also adds tests which demonstrate the passing of these options.