chore: interpolate example code in README.md for each package
This script will modify the given files (all CLI arguments passed to the
script are interpreted as filenames) to interpolate example code from files.

It looks for lines that look like this:

[This example code](../path/to/file.ts) shows how to do X.

And replaces those occurrences to look like this:

This example code shows how to do X.

```ts
// Contents of ../path/to/file.ts
```

This simple hack allows us to keep our code examples in separate files
(which are runnable and testable in CI), while still including them inline in
the documentation that gets published to npmjs.org for the packages.