feat: add usage/fingerprint metadata to `KurtResult` for `KurtOpenAI` adapter
This implements the new optional `metadata` field of the `KurtResult` interface to provide the total
input and output tokens, as well as the system fingerprint string (used for checking whether
determinism can be expected).

Note that this commit involves regenerating all the test snapshots because we are now capturing and
relying on more from the upstream OpenAI responses.