feat: add usage/fingerprint metadata to `KurtResult` for `KurtVertexAI` adapter
This implements the new optional `metadata` field of the `KurtResult` interface to provide the total
input and output tokens. Note that no system fingerprint is available from Vertex AI's Gemini API,
so that is omitted here. If it ever becomes available, we can add it.

Note that this commit involves regenerating all the test snapshots because we are now capturing and
relying on more from the upstream Vertex AI responses.