feat: add `KurtSchema.fromJSONSchema7` utility function
Convert a `JSONSchema7` object schema to a `KurtSchema`, dynamically.

The resulting KurtSchema will be somewhat weakly typed in TypeScript
(the data type is only known to be a `Record` with unknown properties)
because the JSON schema is being dynamically loaded at runtime rather than
baked into the type system statically as Zod types.

However, the underlying LLM will be aware of the actual schema even
if TypeScript isn't, and you'll still get runtime validation of the data.

This function is useful for highly dynamic cases where you need to load
schemas at runtime, from a database, from responses of another service,
or even schemas that were generated by an LLM in a previous step.