fix: kurtCache handling of schemas - serialize as JSON Schema
Prior to this commit, KurtCache would raise an unhandled YAML
serializing exception when storing a cache entry that included
structured data, because it was trying to serialize the Zod
schema object directly. This commit fixes that problem by
converting the schema to JSON schema (which is cleanly serializable)
before writing the YAML cache entry.