fix: correctly pass through the `forceTool` option inside `KurtCache.generateRawEvents`
Prior to this commit, that option was being silently dropped because it was neglected in the new
object that gets sent to the internal adapter. This commit introduces a usage of the spread operator
to ensure that no such options are dropped (while still allowing the replacement of other options).